requirements.txt
streamlit
numerize
pandas
matplotlib
seaborn
plotly-express
tomli
